Force ssl compression always
Disabling ssl compression causes the SSL connection to follow an
entirely different code path, which simply does not work in our
environment. It's unclear what's causing it, but eventlet's read()
method on the SSL connection object throws a SysCallError exception with
errno 9 (Invalid File Descriptor). I've not been able to work out why
that is, so pushing this patch as a temporary measure to work around it.